Add tests for revdep handling in yum update.
This test models the scenario in Bug 773440
regarding clean_requirements_on_remove
(https://bugzilla.redhat.com/show_bug.cgi?id=773440), which
exposes an error in the current implementation of _has_needed_revdeps.

The basic setup is this. For packages A and B, we write A -> B to mean
"B requires A". Packages are dep-installed if they begin with "D" and
user-installed if they begin with "P".

Suppose we have installed {D1, D2, D3, P1} with the relations
D1 -> D2 -> P1 <- D3 and  D3 -> D2. That is, D3 is needed by
both D2 and P1. Suppose we want to update P1 with P1.1 such that
P1.1 no longer needs D3 but still needs D2.

The current remove_old_deps+_has_needed_revdeps combo will incorrectly
try to remove D3 if it hits D3 first when looping through the
deps of P1. This occurs because _has_needed_revdeps does not detect
whether a reverse dependency of D3 is needed by an incoming package;
in this case P1.1 needs D2, which is broken by removing D3.

The patched _has_needed_revdeps checks whether each reverse-dep
is needed by an incoming package. In the above case,
the patched method sees that D2 is needed by P1.1 and marks D2
off-limits for removal. Hence it keeps D3 as well.