Add a metadata_expire_filter configuration, and use it in the shipped commands.
 In theory this is never needed because everyone always runs yum-cron
and repos. are always upto date. In practise everyone complains because
yum hits the network due to that not being the case. This allows the
user to control a lot better how much out of data info. they don't mind
having for simple "yum search" type commands etc.

 Not sure how useful the past/present distinction is, but it's easier to
ignore/remove it that to add it back in later.