Make mdpolicy=group:small default, add group and pkgtags. BZ 989231
With the "metadata_expire_filter" patch from January Yum invalidates
the cache much less often than it used to. This has some pretty bad
consequences when eg users run "yum search" with cache that is few
days old and groups/pkgtags were not loaded yet.

Loading pkgtags and groups with primary should solve this.
(done in parallel with primary and should be cheap).