Mostly backwards compat. change to how distroverpkg config. works. BZ 1002977.
 This is for rel-eng, now instead of just looking at the version of the
package which provides the "distro. release provide" we try to parse the
version out of the packages provides. This allows random verisons for
the package, which is apparently useful.

 Speed of config. loading doesn't seem to be measurable. One change is
that .conf.distroverpkg is now a list to accomodate the new provide vs.
the old one. Also if anyone had a package providing redhat-release = blah,
$releasever is now "blah" instead of whatever the package version was.