Revert "Parse float timestamps as valid, for global timestamp."
This reverts commit e08a0de36ba49075ee9802a9586b49dbd9c3d805.

The float timestamps are assigned to *_db types, and these are
the ones created last, thus having the largest timestamps.

Before this patch, we ignored them, but now they are used and
actually change the global repomd.xml timestamp, that no longer
matches the one from metalink.

We need to keep the old behavior for compatibility.