Allow for validating attributes read from yumdb
Make sure we don't expose corrupted attributes read from the yumdb to
the consumers.  There's at least one report of such a corruption: BZ
1234967.  Instead, make requesting a malformed yumdb attribute
equivalent to requesting a non-existent one -- which is a valid
scenario, already handled by the consumers.

Note that the actual validator function that fixes the above bug will be
committed separately.