yum-cron: don't fail on empty transaction. BZ 1004853
Even if refreshUpdates() returns True, the transaction may still be
empty if some updateinfo filters were applied there and thus a later
call to buildTransaction() would return 0 (success).  This wasn't
handled by findDeps() properly, making it emit an error message in such
a case.

Note that, in the first place, we shouldn't return True from
refreshUpdates() if the transaction becomes empty after applying the
filters.  However, we should handle this particular buildTransaction()
outcome in findDeps() regardless and that's sufficient to fix this bug.

See also:
http://lists.baseurl.org/pipermail/yum/2014-December/024141.html