updateinfo: accept respin in advisory name. BZ 1065380
When addressing Red Hat errata, we don't expect nor want the users to
care about respin numbers.  Still, in some places like the RHN portal,
we do display them in the form of an advisory name suffix.  Since the
updateinfo data usually doesn't include them either, we currently return
no matches if the user specifies such an extended form on the cmdline
(e.g. by copy-pasting it from the RHN portal).

This commit addresses that by either interpreting such an advisory name
exactly as specified or, if that would yield zero matches, taking it as
if the respin suffix was omitted.  This applies to all such places where
advisory names are expected on the cmdline.

Examples of new behavior:

    # updateinfo only contains RHSA-2016:1234 (typical scenario)
    yum updateinfo RHSA-2014:1234-3  # matches RHSA-2016:1234
    yum updateinfo RHSA-2014:1234    # matches RHSA-2016:1234

    # updateinfo also contains RHSA-2016:1234-3
    yum updateinfo RHSA-2014:1234-3  # matches RHSA-2016:1234-3 (only)
    yum updateinfo RHSA-2014:1234    # matches RHSA-2016:1234 (only)