Handle lvm exceptions from fssnap. BZ 1335250
Currently, we don't handle the LibLVMError that may happen during the
LVM API calls, so when it occurs, we just crash.  That's unfortunate
because:

- Only an unfriendly traceback is printed

- fssnap_abort_on_errors is not honored in such cases (transaction
  aborts even if "none" is set)

This commit improves that in the following ways:

- Print a user-friendly error message instead of a traceback

- Never abort a transaction on fssnap errors unless configured via
  fssnap_abort_on_errors

- If configured to abort a transaction, do it in the standard way by
  raising the YumRPMTransError

- Properly treat empty (aka erroneous) return values from LVM calls by
  raising an exception instead of just silently returning

- Defer the computation of VG names until needed, so that we don't have
  to handle any exceptions during the _FSSnap construction

Note:

There's no documentation for the LVM python binding that would tell us
what exceptions we may anticipate.  From what I saw when inspecting the
binding's source code, though, the LibLVMError seems to be the only one
that is used to indicate errors during LVM operations, which is what we
are interested in catching.