Fix reget problems with simple md filenames. BZ 1155687
Whenever we move to a new repomd version by either preloading or
downloading it, we first look whether we already have the metadata files
that it requires so that we avoid redownloading the same stuff.  This is
done by testing for the presence of filenames recorded in the new repomd
and comparing their sizes to those we expect.  Those that don't exist we
redownload and those that do but are smaller we try to reget.

This logic works fine for unique filenames because they always belong to
one specific repomd only.  However, for simple filenames, it usually
gives false positives whenever the new repomd is different from the old
one, making us produce corrupt files as a result.  This commit fixes
that by doing a full download in such cases.

Pseudocode equivalent of the original (unfixed) logic:

   def preload(file):
       copy_from_system_cache(file) if not running_as_root and
                                       age_system(file) < age_user(file)

1  preload(repomd)
2  if expired:
3      download(repomd)
4      for file in repomd.files:
5          preload(file)
6          done if repomd.valid(file)
7          reget(file) if size(file) < repomd.size(file)
8                      else download(file)
9  else:
10     for file in repomd.files:
11         preload(file) if not exists(file)
12         done if repomd.valid(file)
13         reget(file) if size(file) < repomd.size(file)
14                     else download(file)

Common scenarios that would lead to the problem:

ROOT | REPO | SYSTEM | USER | EXP | SIZE   | PATH        | OUTCOME
-------------------------------------------------------------------
yes  | B/B  | A/A*   | -/-  | yes | A* < B | 2-3-4-7     | corrupt
no   | C/C  | B/B    | A/A  | yes | B < C  | 1-2-3-4-5-7 | corrupt
no   | B/B  | B/B    | A/A  | no  | A < B  | 1-2-9-10-13 | corrupt
no   | C/C  | B/B    | A/A  | no  | A >= B | 1-2-9-10-14 | incompat

Legend:
ROOT = running as root?
REPO = metadata version (repomd/file) in the repo
SYSTEM = metadata version (repomd/file) in the system cache
USER = metadata version (repomd/file) in the user cache
EXP = is the repo expired?
SIZE = file size comparison
PATH = path in the pseudocode snippet above
OUTCOME = what happens to the file as a result (incompat means it
          doesn't belong to the current repomd)
X* = partially downloaded file X