sqlitesack: fix filelist queries for duplicate pkgs. BZ 1174380
Make sure we don't return duplicate file/changelog entries for packages
that have some duplicates in the DB.

While we don't exactly support duplicate packages in the code, there was
a risk someone would (even unwillingly) maintain a repo with such dupes,
in which case createrepo would grow the filelist size each time it was
run with --update, in a similar fashion to BZ 1174380.  This commit
prevents that.