Detect installed virtual provide in install(). BZ 1352585
Normally, when the user tries to install something that's already
installed, we exit gracefully with 0.  However, that's not the case if
what we're looking for is a provide that, despite being installed, is
not available in any of the enabled repos, in which case we error out.
This commit makes sure we exit gracefully in that case too.

The old code path for "yum install foo" looks like this:

1) Look for foo in pkgSack
2) If no success, look for a package in pkgSack providing foo
3) If no success, look for foo in rpmdb
4) If no success, error out with "No package foo available." and exit
   code 1

What we're adding with this commit is the following in between 3 and 4:

- If no success, look for a package in rpmdb providing foo

Note that we only search for the provide in pkgSack if the kwarg
'pattern' is set, so let's adhere to this with the newly added rpmdb
search too.