Port to the official gpgme python bindings
Mostly adapted from DNF commit 5145f588.

Note that we're removing the piece of code that checked the validation
result in valid_detached_sig(); this is now done by gpg.Context.verify()
itself which raises an exception if it fails.

This method also doesn't require the key to be trusted in order to use
it for verification so we no longer have to trust the key after
importing it (which is not possible anyway since the new bindings don't
support adjusting key trust at this time).