Don't pollute stderr in FIPS mode. BZ 1287610
When the system is in FIPS mode, certain unsecure checksum algorithms
such as md5 are disabled by hashlib and thus not available in yum.  We
normally print an error during misc.py initialization for every such
algorithm, however that's just too annoying if it happens on every yum
run and is also confusing (as it doesn't say why the particular
algorithm is disabled).

Let's stop printing the message if the cause is FIPS mode and also
extend the message that is printed when yum actually tries to use such a
checksum algorithm (e.g. during repodata verification of a repo using
md5) so that it says it's due to FIPS.

Note: Custom FIPSNonCompliantError(MiscError) class is used instead of
just MiscError so that we can make the resulting message shorter in
YumRepository._checksum() without changing the message for any other
exceptions handled in that method.