depsolve: filter out conflicting provider. BZ 1480065
When there are multiple providers available for a requirement, yum would
happily pick the one that the requiring package also has a "Conflicts:"
on (via another virtual provide), failing to resolve the transaction.

Example:
- foo requires bar and conflicts my-bar
- bax provides bar
- bay provides bar, my-bar

Yum might decide to pick bay, only to fail due to the conflict with foo
later in the process.

This commit fixes that by dropping such a provider from the candidate
list when depsolving.