Print remaining cache size on "yum clean all". BZ 1481220
Drop the unpopular "rm -rf" hint and replace it with actual info on how
much data from other repos, if any, still remains in the cache.  This
serves two purposes:

    - Remind the user that "yum clean all" doesn't mean "purge all the
      things"
    - Refer the user to --verbose mode to learn why that is
    - Refer the user to the appropriate man page section if the above
      doesn't help

Note the focus on being as succinct as possible.  We don't want to
pollute the output necessarily on every "yum clean all".