installonlypkgs: do not reinstall local pkgs. BZ 1689025
This commit fixes the following scenario:

$ rpm -q kernel
kernel-1
kernel-2
kernel-3
$ yum install /path/to/kernel-2.rpm
[...]
Reinstalling:
 kernel
[...]

The root case was that we would put the new kernel in the updatepkgs
list if there was an older kernel in the rpmdb, resulting in a reinstall
action.  To fix this, let's make sure we put it in installpkgs where it
belongs.  This makes yum just skip the installed kernel when doing a
local install, just like the regular install.

Note:  Ideally, we would simply move the other allowedMultipleInstalls()
stub in front of the loop to fix this.  We can't do that, however, since
exactarchlist would no longer be honored, causing a regression (even
though that would actually be correct and consistent with the regular
install command).