installonlypkgs: do not reinstall when local. BZ 1689025
Previously, when installing a local package that is listed in
installonlypkgs, we would try to reinstall it if there was an older
version in the rpmdb already.  For example:

$ rpm -q kernel
kernel-1
kernel-2
kernel-3
$ yum install /path/to/kernel-2.rpm
[...]
Reinstalling:
 kernel
[...]

This was because we would end up populating the updatepkgs list in
installLocal() due to kernel-1 being present, which would then translate
into a reinstall action.

In addition to that, we would not pass the appropriate "please
reinstall" RPM flags, making the transaction fail due to kernel-2 being
present.

To fix this, we could either pass those RPM flags and make it really
reinstall the package, or, better yet, just skip the package.  The
latter is actually more sane and consistent with the regular (non-local)
install command, so that's what this commit accomplishes.