Gracefully handle self-conflicts in RPM. BZ 1564747
Self-conflicts are normally fine and supported by both RPM and YUM,
however there's one code path (reinstall) which is currently not handled
in RPM and causes us to fail.

This commit adds a workaround that enables reinstalls to work properly
when the package has self-conflicts in the rpmdb already.

In the long term, this should be fixed on the RPM side.