installonlypkgs: install packages related to the currently running kernel in the docker container
When we are in a docker container and the kernel currently running is:
$ uname -a
Linux 5001cc62a821 4.19.57-15.1.al7.x86_64 #1 SMP Thu Aug 29 13:46:41
CST 2019 x86_64 x86_64 x86_64 GNU/Linux

Then we need to install the kernel-headers package for some scenarios.
At that time, the yum command automatically obtained the latest package,
not the one related to the current kernel：
$ sudo yum install kernel-headers
Loaded plugins: branch, fastestmirror, langpacks
Loading mirror speeds from cached hostfile
Resolving Dependencies
--> Running transaction check
---> Package kernel-headers.x86_64 0:4.19.67-16.al7 will be installed
...

As we are only in a container, so we cannot upgrade the host's kernel,
and we can only install packages consistent with the host's kernel.

Signed-off-by: Wen Yang <wenyang@linux.alibaba.com>
Cc: Caspar Zhang <caspar@linux.alibaba.com>